% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yields.R
\name{get_yields}
\alias{get_yields}
\title{get_yields}
\usage{
get_yields(series = "DGS10", na_locf = T, percent_adjust = T, ...)
}
\arguments{
\item{series}{The series code as found on https://fred.stlouisfed.org/}

\item{...}{Additional parameters handed to quantmod::getSymbols}
}
\value{
The yields data as xts
}
\description{
get_yields
}
\details{
1-Month Treasury Constant Maturity Rate: DGS1MO
10-Year Treasury Constant Maturity Rate: DGS10
}
