% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_return.R
\name{total_return}
\alias{total_return}
\title{Title}
\usage{
total_return(
  yields,
  maturity,
  mdur = mod_duration(yields, maturity),
  convex = convexity(yields, maturity),
  scale = 261,
  format_out = "xts"
)
}
\arguments{
\item{format_out}{}
}
\value{

}
\description{
Title
}
