% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexity.R
\name{convexity}
\alias{convexity}
\title{Calculate the convexity of a bond}
\usage{
convexity(yields, maturity, format_out = "xts")
}
\arguments{
\item{yields}{a series of yields}

\item{maturity}{constant bond maturity in years}

\item{format_out}{xts or tibble}
}
\value{

}
\description{
Calculates the convexity of a bond.

This function is normally used in combination with total_return() to compute bond total returns.
}
\details{
This function is normally used in combination with total_return() to compute bond total returns.

The convexity is the interest rate sensitivity of the modified duration.

\usepackage{lmodern}

\deqn{convexity = C_1 - C_2}{hello}

where

\deqn{C_1 = \frac{2}{y_t^2} (1-{z_t}^{-2M})}{hello}
\deqn{C_2 = \frac{2M}{y_t}{z_t}^{-2M-1}}{hello}
\deqn{z_t = 1+\frac{y_t}{2}}{hello}

\eqn{M}{M} is the maturity in years (e.g. 10), \eqn{y_t}{t_y} is the yield at time \eqn{t}{t}.
}
