% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_to_tibble.R
\name{xts_to_tibble}
\alias{xts_to_tibble}
\title{Convert an xts object to a tibble data frame}
\usage{
xts_to_tibble(x)
}
\arguments{
\item{x}{an xts object}
}
\value{
A tibble data frame with first column 'date'
}
\description{
Convert an xts object to a tibble data frame
}
